Split `build` and `host`
In conda-build 3, the new `host` environment was created and split
from `build` requirements. If both are listed, conda-build creates
two separate environments each with their own requirements. The build
environment is only used to provide command line tools (not libraries)
for the build and the host environment is for all dependencies that are
linked to the package.

Here we split the build and host environments and move `cmake` into
the `build` requirements. This should ensure the package is not linked
to `openssl` at run time.